Redirect back to schedule after requesting assignment.
Before the signup form would redirect back to the form.  The prompts changed slightly but it wasn't really clear to the casual user that the assignment request worked.

With ths patch the user is redirect back to the schedule where they can see their name assigned to the game.  Think it is more clear.